Backend_new/requirements.txt
Django>=5.2.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

